+++ src/Traits/CollectionTrait.php
@@ -8,10 +8,14 @@
 
 namespace Meritoo\Common\Traits;
 
+use Meritoo\Common\Traits\Collection\AddTrait;
 use Meritoo\Common\Traits\Collection\ArrayAccessTrait;
 use Meritoo\Common\Traits\Collection\CountableTrait;
+use Meritoo\Common\Traits\Collection\GetTrait;
 use Meritoo\Common\Traits\Collection\IteratorAggregateTrait;
 use Meritoo\Common\Traits\Collection\MainTrait;
+use Meritoo\Common\Traits\Collection\ModifyTrait;
+use Meritoo\Common\Traits\Collection\VerifyTrait;
 
 /**
  * Trait for the Collection
@@ -22,6 +26,10 @@ use Meritoo\Common\Traits\Collection\MainTrait;
 trait CollectionTrait
 {
     use MainTrait;
+    use AddTrait;
+    use ModifyTrait;
+    use GetTrait;
+    use VerifyTrait;
     use CountableTrait;
     use ArrayAccessTrait;
     use IteratorAggregateTrait;